pyproject.toml
[tool.poetry]
name = "smart-dotenv"
version = "1.0.0"
description = "A minimal utility to load .env variables with automatic type casting."
authors = ["Antares Mugisho <antaresmugisho@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
